+++ Modules/TradeUI.lua
@@ -163,6 +163,11 @@ function TradeUI:OnEvent_UI_INFO_MESSAGE (event, ...)
    if select(1, ...) == _G.LE_GAME_ERR_TRADE_COMPLETE then -- Trade complete. Remove items from db.baggedItems if traded to winners
       addon:Debug("TradeUI: Traded item(s) to", self.tradeTarget)
       for _, link in ipairs(self.tradeItems) do
+         local Item = addon.ItemStorage:GetItem(link)
+         if Item and addon:UnitIsUnit(self.tradeTarget, Item.args.recipient) then
+            -- REVIEW This check is probably redundant, but currently TRADE_ACCEPT_UPDATE just adds all traded items, not just those we track
+            addon:SendCommand("group", "trade_complete", link, self.tradeTarget, addon.playerName)
+         end
          addon.ItemStorage:RemoveItem(link)
       end
       self:Update()